Don't hide the new comment form on a fatal error
It's possible that the user entered an invalid OAuth token, and now you've removed the only way for them to fix it.
When that happens, null out the OAuth token to prevent further fatal errors from bad credentials.
When a comment is posted, show all of the comments if they're aren't shown already.
Move fatal error div below new comment form.

TODO:
 * Guidance, reassurance about tokens/authorization
 * Skinning